+++ interfaces/system_key_test.go
@@ -36,7 +36,6 @@ type systemKeySuite struct {
 	tmp              string
 	apparmorFeatures string
 	buildID          string
-	restorers        []func()
 }
 
 var _ = Suite(&systemKeySuite{})
@@ -49,20 +48,16 @@ func (s *systemKeySuite) SetUpTest(c *C) {
 	id, err := osutil.MyBuildID()
 	c.Assert(err, IsNil)
 	s.buildID = id
-
-	s.restorers = []func(){
-		osutil.MockMountInfo(""), osutil.MockEtcFstab(""),
-	}
 }
 
 func (s *systemKeySuite) TearDownTest(c *C) {
 	dirs.SetRootDir("/")
-	for _, fn := range s.restorers {
-		fn()
-	}
 }
 
 func (s *systemKeySuite) TestInterfaceSystemKey(c *C) {
+	restore := interfaces.MockIsHomeUsingNFS(func() (bool, error) { return false, nil })
+	defer restore()
+
 	systemKey := interfaces.SystemKey()
 
 	apparmorFeatures := release.AppArmorFeatures()
@@ -85,6 +80,8 @@ apparmor-features:
 - caps
 - dbus
 `)
+	defer restore()
+	restore = interfaces.MockIsHomeUsingNFS(func() (bool, error) { return false, nil })
 	defer restore()
 
 	systemKey := interfaces.SystemKey()
